backend/requirements.txt
asgiref==3.9.2
bcrypt==5.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.7
djangorestframework==3.16.1
drf-yasg==1.21.11
idna==3.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
